Change unsigned long to size_t everywhere. Remove Allman.cpp (no longer of interest).

+++ GenerateBoundaryPairs/GeneratePairsOfTestData.h
@@ -11,25 +11,25 @@
 class GeneratePairsOfTestData {
 public:
    static std::pair<S6, S6> ScalePairToSpecifedSeparation(const S6& s1, const S6& s2, const double factor);
-   static void GenerateRandomLatticesAndCatalogByReductionStepCount(const unsigned long n);
-   static std::pair<S6, S6> GenerateLatticeTypeExamplesNearDeloneTypes(const unsigned long n);
+   static void GenerateRandomLatticesAndCatalogByReductionStepCount(const size_t n);
+   static std::pair<S6, S6> GenerateLatticeTypeExamplesNearDeloneTypes(const size_t n);
    static S6 DoRandomReflection(const S6& s);
 
 
-   static void MultiGeneratePair_Reduced_Other_ReducesFarAway(const unsigned long targetReducedUnreducedPairCount);
+   static void MultiGeneratePair_Reduced_Other_ReducesFarAway(const size_t targetReducedUnreducedPairCount);
    static double PositiveDistance(const S6& s6);
    static std::string FormatPairOfTestS6Vectors(const S6& s1, const S6& s2, const std::string& id);
 private:
-   static bool IsGoodLatticeTypeForPerturbation(const unsigned long n);
+   static bool IsGoodLatticeTypeForPerturbation(const size_t n);
    static std::pair<S6, S6> OneLatticeType(const MatS6& m);
    static std::pair<S6, S6> OneLatticeAttempt(const MatS6& m);
-   static std::pair<S6, S6> GeneratePair_Reduced_Other_ReducesFarAway(const unsigned long numberOfZeros);
-   static S6 DoReflection(const S6& s, const unsigned long n);
+   static std::pair<S6, S6> GeneratePair_Reduced_Other_ReducesFarAway(const size_t numberOfZeros);
+   static S6 DoReflection(const S6& s, const size_t n);
    static std::pair<S6, S6> BinarySearchReduced(const S6& s1, const S6& s2, const int npass);
    static S6 GenerateStartingPoint(const MatS6& m);
    static int Rank(const S6& s);
    static S6 Perturb(const S6& s, const double fraction);
-   static void StoreOnePointIfValid(StoreResults<unsigned long, std::string >& store, const S6& s6in);
+   static void StoreOnePointIfValid(StoreResults<size_t, std::string >& store, const S6& s6in);
 };
 
 #endif  //  GENERATEPAIRSOFTESTDATA_H
